Move models to model.py
